+++ src/azure_devtools/scenario_tests/preparers.py
@@ -4,11 +4,10 @@
 # --------------------------------------------------------------------------------------------
 
 import contextlib
-import inspect
 import functools
 
 from .base import ReplayableTest
-from .utilities import create_random_name, is_text_payload
+from .utilities import create_random_name, is_text_payload, trim_kwargs_from_test_function
 from .recording_processors import RecordingProcessor
 
 
@@ -48,7 +47,7 @@ class AbstractPreparer(object):
             if parameter_update:
                 kwargs.update(parameter_update)
 
-            _trim_kwargs_from_test_function(fn, kwargs)
+            trim_kwargs_from_test_function(fn, kwargs)
 
             fn(test_class_instance, **kwargs)
 
@@ -120,41 +119,3 @@ class SingleValueReplacer(RecordingProcessor):
         self.replace_header(response, 'azure-asyncoperation', self.random_name, self.moniker)
 
         return response
-
-
-class AllowLargeResponse(object):  # pylint: disable=too-few-public-methods
-
-    def __init__(self, size_kb=1024):
-        self.size_kb = size_kb
-
-    def __call__(self, fn):
-        def _preparer_wrapper(test_class_instance, **kwargs):
-            from azure_devtools.scenario_tests import LargeResponseBodyProcessor
-            large_resp_body = next((r for r in test_class_instance.recording_processors
-                                    if isinstance(r, LargeResponseBodyProcessor)), None)
-            if large_resp_body:
-                large_resp_body._max_response_body = self.size_kb  # pylint: disable=protected-access
-
-            _trim_kwargs_from_test_function(fn, kwargs)
-
-            fn(test_class_instance, **kwargs)
-
-        setattr(_preparer_wrapper, '__is_preparer', True)
-        functools.update_wrapper(_preparer_wrapper, fn)
-        return _preparer_wrapper
-
-# Utility
-
-def _trim_kwargs_from_test_function(fn, kwargs):
-    # the next function is the actual test function. the kwargs need to be trimmed so
-    # that parameters which are not required will not be passed to it.
-    if not is_preparer_func(fn):
-        args, _, kw, _ = inspect.getargspec(fn)  # pylint: disable=deprecated-method
-        if kw is None:
-            args = set(args)
-            for key in [k for k in kwargs if k not in args]:
-                del kwargs[key]
-
-
-def is_preparer_func(fn):
-    return getattr(fn, '__is_preparer', False)
